Refactor : getElapsedTime 함수 분리 및 매직 넘버 상수화
